Properly stop characters crossing over in an event
By defining a global sort order for which threads come first, rather
than relying on internal event orderings that might not be consistent.

I did this to make the code compatible with propagation (next commit),
which changed characters within events to sets (so I couldn't rely on
that order any more), but it turns out it's a pretty decent improvement
on its own.