Break out non-working layout code to own folder
As a step towards cleanup for Code Climate. I'm not quite game to fully
remove this yet, but at least if it's in its own folder and nothing else
depends on it I can remove it cleanly in future. And in the meantime
it's easier to focus on the other Code Climate warnings that I care more
about.