Fix weird Code Climate warning re methods in []
I'm not sure why Code Climate expects these method calls to be aligned
with the square brackets around them, and the documentation doesn't
really explain it.

But extracting a variable is simple and should presumably fix the
warning (modulo maybe a couple of tweaks to get to their preferred
indentation rule).